migrate tests to ava
