+++ .babelrc
@@ -1,0 +1,12 @@
+{
+  "presets": [
+    ["@babel/preset-env", {
+      "targets": {
+        "node": "4"
+      },
+      "exclude": [
+        "transform-regenerator"
+      ]
+    }]
+  ]
+}
